# public/dist/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Demo App",
    "description": "",
    "contact": {
        "email": ""
    },
    "version": "1.3.0"
  },

  "servers":[
    {
      "url":"http://localhost:5001/demo/api/v2",
      "description":"Local url"
    }
  ],

  
  "consumes": [
    "application/json",
    "application/x-www-form-urlencoded",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
 
  "paths": {
    "/user/login": {
      "post": {
        "deprecated":true,        
        "tags": [
          "User"
        ],
        "description": "Login",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
            "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "formData",
            "description": "email",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "password",
            "required": true,
            "type": "string"
          },
          {
            "name": "deviceToken",
            "in": "formData",
            "description": "deviceToken",
            "required": true,
            "type": "string"
          },
          {
            "name": "deviceTypeID",
            "in": "formData",
            "description": "1= IOS, 2=Android , 3=web",
            "required": true,
            "enum":[1,2,3],
            "type": "integer"
          },
          {
            "name": "currentVersion",
            "in": "formData",
            "description": "currentVersion",
            "required": true,            
            "type": "string"
          }
        ]
      }
    },
    "/user/signup": {
      "post": {
        "deprecated":true,        
        "tags": [
          "User"
        ],
        "description": "Signup",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
            "application/json"
        ],
        "parameters": [
          {
            "name": "deviceID",
            "in": "formData",
            "description": "deviceID",
            "required": true,
            "type": "string"
          },
          {
            "name": "deviceToken",
            "in": "formData",
            "description": "deviceToken",
            "required": true,
            "type": "string"
          },
          {
            "name": "deviceTypeID",
            "in": "formData",
            "description": "1= IOS, 2=Android , 3=web",
            "required": true,
            "enum":[1,2,3],
            "type": "integer"
          },
           {
            "name": "email",
            "in": "formData",
            "description": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "description": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "dob",
            "in": "formData",
            "description": "date of birth - YYYY-MM-DD",
            "type": "string"
          },
          {
            "name": "gender",
            "in": "formData",
            "description": "0 Undefined, 1 Male, 2 Female, 3 Others",
            "enum":[0,1,2,3],
            "type": "integer"
          },
          {
            "name": "currentVersion",
            "in": "formData",
            "description": "currentVersion",
            "type": "string",
            "required": true            
          }
         
         
        ]
      }
    },
    "/user/logout": {
      "post": {
        "deprecated":true,        
        "tags": [
          "User"
        ],
        "description": "Social / normal login - logout",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
            "application/json"
        ],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "accessToken",
            "required": true,
            "type": "string"
          }
        ]
      }
    },


    "/user/updateProfile": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Update Profile",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
            "application/json"
        ],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "accessToken",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "name of user",
            "required": true,
            "type": "string"
          },
          {
            "name": "username",
            "in": "formData",
            "description": "userName of the user",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
            "200": {
                "description": "Successfully"
            },
            "401": {
                "description": "Unauthorized"
            }
        },
        "security": [
            {
                "basicAuth": [
                    "Username: ",
                    "Password: "
                ]
            }
        ]
      }
    },


    "/user/social-login": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Social Signup / Login",
        "consumes": [
            "application/x-www-form-urlencoded"
        ],
        "produces": [
            "application/json"
        ],
        "parameters": [
          {
            "name": "deviceID",
            "in": "formData",
            "description": "deviceID",
            "required": true,
            "type": "string"
          },
          {
            "name": "deviceToken",
            "in": "formData",
            "description": "deviceToken",
            "required": true,
            "type": "string"
          },
          {
            "name": "deviceTypeID",
            "in": "formData",
            "description": "1= IOS, 2=Android , 3=web",
            "required": true,
            "enum":[1,2,3],
            "type": "integer"
          },
          {
            "name": "socialType",
            "in": "formData",
            "description": "1= Facebook, 2=Google",
            "required": true,
            "enum":[1,2],
            "type": "integer"
          },
          {
            "name": "socialId",
            "in": "formData",
            "description": "",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "description": "email",
            "type": "string"
          },
          {
            "name": "dob",
            "in": "formData",
            "description": "date of birth - YYYY-MM-DD",
            "type": "string"
          },
          {
            "name": "name",
            "in": "formData",
            "description": "name",
            "type": "string"           
          },
          {
            "name": "username",
            "in": "formData",
            "description": "username",
            "type": "string",
            "required": true     
          },
          {
            "name": "gender",
            "in": "formData",
            "description": "0 Undefined, 1 Male, 2 Female, 3 Others",
            "enum":[0,1,2,3],
            "type": "integer"
          },
          {
            "name": "profileImage",
            "in": "formData",
            "description": "profileImage",
            "type": "string"   
          }
        ],
        "responses": {
            "200": {
                "description": "Successfully"
            },
            "401": {
                "description": "Unauthorized"
            }
        },
        "security": [
            {
                "basicAuth": [
                    "Username: ",
                    "Password: "
                ]
            }
        ]
      }
    },
    "/user/guest-login": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Guest Login",
        
        "produces": [
            "application/json"
        ],
        "requestBody":{
                "content":{
                  "multipart/form-data":{
                    "schema":{
                      "$ref":"#/components/schemas/ImageFile"
                    }
                  }
                }
            },
        "responses": {
            "200": {
                "description": "Successfully"
            },
            "401": {
                "description": "Unauthorized"
            }
        },
        "security": [
            {
                "basicAuth": [
                    "Username: ",
                    "Password: "
                ]
            }
        ]
      }
    },

    "/user/getProfile": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "GET Profile",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
            "application/json"
        ],
        "parameters": [
          {
            "name": "accessToken",
            "in": "header",
            "description": "accessToken",
            "required": true,
            "type": "string"
          }
        ],
       
        "responses": {
            "200": {
                "description": "Successfully"
            },
            "401": {
                "description": "Unauthorized"
            }
        },
        "security": [
            {
                "basicAuth": [
                    "Username: ",
                    "Password: "
                ]
            }
        ]
      }
    }
  },

  "securityDefinitions": {
        "basicAuth": {
            "type": "basic"
        }
    },

    "components":{
      "schemas":{
        "ImageFile":{
            "type":"object",
            "properties":{
              "profileImage":{
                "type":"string",
                "format":"binary",
                "required":true
              },
              "deviceID":{
                "description": "deviceID",
                "required": true,
                "type": "string"
              },
              "deviceToken":{
                "description": "deviceToken",
                "required": true,
                "type": "string"
              },
              "deviceTypeID":{
                "description": "1= IOS, 2=Android , 3=web",
                "required": true,
                "enum":[1,2,3],
                "type": "integer"
              },
              "name":{
                "description": "name",
                "type": "string",
                "required":true           
              },
              "username":
              {
                "description": "username",
                "type": "string",
                "required": true     
              }
            }
        }
      }
    }
}